Add fact for installed software
This commit adds a custom fact that checks for the presence of the
specified packages.  Currently, that's "sudo" and "tcp_wrappers".

It simply uses 'rpm' to check the status and will return "Failed" if
either of them are not present.

It may be desirable to break these out into separate facts for each
package(?), but this replicates the current implementation.