Add fact to check ftp logging compliance
This commit adds a custom fact that verifies FTP services on
el(2|[4-7]).  It uses Augeas to verify.

On EL5 and 6, this verifies vsftpd via the vsftpd.conf file.

On EL4, this verifies gssftp via the xinetd.d file for the service.

On EL 2 and 7, this verifies wu-ftpd via the xinetd.d file for the
service.

If the service is NOT enabled, this will return "passed", as it's
considered compliant.

If the service is enabled, logging must be enabled.