Add puppet class to enforce ftp compliance
This commit adds a Puppet class to enforce ftp logging compliance.

This uses Augeas to ensure configuration.

On EL5 and 6, this alters vsftpd via the vsftpd.conf file.
On EL4, this configures gssftp via the xinetd.d file for the service.

On EL 2 and 7, this configures wu-ftpd via the xinetd.d file for the
service.