Add fact to verify umask
This commit adds a custom fact to verify 'umask' settings in specified
files.  Currently, that's '/etc/profile' and '/etc/bashrc'

It's naive.  It simply searches for lines that set a 'umask' that
doesn't match a specified umask.  If any inconsistent lines are found,
this fact will report "Failed."

It will not search sourced files from the files it checks.

Reference:
http://www.stigviewer.com/stig/red_hat_enterprise_linux_5/2014-04-02/finding/V-808

This indicates that umasks should be 077.  This fact uses 022 by
default.