Add non-compliant accounts to list more
Prior to this commit, accounts that were COMPLIANT on RHEL 2,3,4,7 were
added to the 'troubled_accounts' list, while non-compliant accounts were
not.

This commit reverses that - adding the non-compliant system accounts to
the troubled_accounts array so that they can be reported by the
'fix_disable_accounts' fact.

Also adjust indentation on a block