Set a default status for check_syslog
The check_syslog fact was returning the value of 'status', however, that
may not be set.  This defaults it to 'Failed' and an explicit pass can
change it.

If '/etc/syslog.conf' doesn't exist (it doesn't on CentOS 6.5), the
'status' never got set because Augeas couldn't evaluate the file,
resulting in an empty fact.