Fix logic in check_user_policy; Add fix fact
Prior to this commit, the 'check_user_policy' fact was improperly
checking for the existence of an Augeas path, resulting in false
positives.  It was checking only part of the context.  This commit
changes that behavior to check for the existence of the absolute path
(full context).

The third policy check against PAM files was also improperly assigning
the 'check' variable (== vs =).

This commit also adds a 'fix_user_policy' fact, which will provide a
comma-separated list of the files that contained non-compliant entries.

This has been tested for compliance and non-compliance on a CentOS 6
system.